Rename examples steps example components to kebab-case
