--- packages/examples/src/components/examples/StepsExample/index.ts
@@ -1,6 +1,0 @@
-export * from './CredentialsStep';
-export * from './FinishStep';
-export * from './InterestsStep';
-export * from './ProfileStep';
-export * from './VerificationStep';
-export * from './WelcomeStep';
