+++ packages/examples/src/components/examples/steps/index.ts
@@ -1,0 +1,6 @@
+export * from './credentials-step';
+export * from './finish-step';
+export * from './interests-step';
+export * from './profile-step';
+export * from './verification-step';
+export * from './welcome-step';
